# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IndianTradeMart API
  version: 1.0.0
  description: Express API surface for the IndianTradeMart platform.
servers:
  - url: http://localhost:3001
    description: Local dev server
  - url: / 
    description: Relative (proxy)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    KycReject:
      type: object
      properties:
        remarks:
          type: string
...
paths:
...
  /api/kyc/vendors/{vendorId}/documents:
    get:
      summary: Fetch vendor KYC documents
      tags: [KYC]
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Documents
  /api/kyc/vendors/{vendorId}/approve:
    post:
      summary: Approve vendor KYC
      tags: [KYC]
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved
  /api/kyc/vendors/{vendorId}/reject:
    post:
      summary: Reject vendor KYC
      tags: [KYC]
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycReject'
      responses:
        '200':
          description: Rejected
...
  /api/finance/payments:
    get:
      summary: Finance payment list
      tags: [Finance]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Payments
  /api/finance/summary:
    get:
      summary: Finance summary
      tags: [Finance]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Summary
  /api/finance/coupons:
    get:
      summary: List coupons
      tags: [Finance]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Coupons
    post:
      summary: Create coupon
      tags: [Finance]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, discount_type, value]
              properties:
                code:
                  type: string
                discount_type:
                  type: string
                value:
                  type: number
      responses:
        '200':
          description: Created
  /api/finance/coupons/{code}/deactivate:
    post:
      summary: Deactivate coupon
      tags: [Finance]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deactivated
  /api/finance/coupons/{id}:
    delete:
      summary: Delete coupon
      tags: [Finance]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...
  /api/superadmin/finance/summary:
    get:
      summary: Superadmin finance summary
      tags: [Superadmin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Summary
  /api/superadmin/finance/payments:
    get:
      summary: Superadmin finance payments
      tags: [Superadmin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Payments
...